refactor(tests): consolidate test files into source-to-test 1:1 companions
Cleans up the source-to-test 1:1 violation introduced by PRs #412
and #413 — both shipped extra test files whose contents belonged
in the existing companion *_test.go files for the source files
they targeted. Pure mechanical refactor; zero test logic changed,
test count preserved exactly (22 == 22 across the affected files,
41 == 41 across all touched packages including pre-existing tests).

Files:
- DELETED database/internal/columns/metadata_extract_test.go;
  13 tests moved to metadata_test.go (already the 1:1 companion
  for metadata.go where every test target lived).
- DELETED database/types/uncovered_branches_test.go; its 9 tests
  redistributed by target-source:
  - 6 tests targeting interfaces.go (NewRowFromSQL,
    sqlRowAdapter.Scan/Err) → new database/types/interfaces_test.go
  - 2 tests targeting subquery.go (ValidateSubquery validator-
    interface fast-path) + the mockValidatingSubquery helper
    appended to subquery_test.go alongside the existing
    mockValidSubquery it embeds.
  - 1 test targeting table.go (TableRef.As nil-receiver) appended
    to table_test.go.

Why: the repo convention is exactly one *_test.go companion per
source file. The violating files used naming-as-changelog
(`*_extract_test.go`, `*_uncovered_branches_test.go`) which mixes
"the gap-closing batch" semantics into file structure that should
encode "which source file does this test target". The convention
is now memory-pinned (feedback_source_to_test_1to1.md) so future
gap-closing tests will land in the right file the first time.

Pre-push: /simplify APPROVE (no helper duplication, imports clean,
section-divider comments preserved, mockValidatingSubquery's
embedding of mockValidSubquery is valid since both now co-reside).
/security-audit PASS (test-count preserved at 41==41, panic
assertions intact verbatim — no require→assert downgrade,
interfaces_test.go correctly retains `package types`, no production
code changes).

Refs feedback_source_to_test_1to1.md memory entry.
