+++ database/types/interfaces_test.go
@@ -1,3 +1,4 @@
+//revive:disable-next-line:var-naming // Package name "types" avoids circular imports.
 package types
 
 import (
@@ -73,47 +74,4 @@ func TestSqlRowAdapterErrDelegatesToUnderlyingSQLRow(t *testing.T) {
 
 	assert.Error(t, wrapped.Err(), "deferred query error must surface through Err() delegation")
 	require.NoError(t, mock.ExpectationsWereMet())
-}
-
-// ---------------------------------------------------------------------------
-// ValidateSubquery — covers the previously-uncovered subqueryValidator
-// fast-path branches (success and error). Existing tests cover the
-// nil/invalid-ToSQL/empty-SQL paths via plain mocks; here we use mocks
-// that ALSO implement subqueryValidator to exercise the early-return
-// code path before ToSQL is even called.
-// ---------------------------------------------------------------------------
-
-// mockValidatingSubquery implements both SelectQueryBuilder and the unexported
-// subqueryValidator interface. ValidateForSubquery is wired by the caller.
-type mockValidatingSubquery struct {
-	mockValidSubquery // reuse the existing valid-mock for the SelectQueryBuilder surface
-	validateErr       error
-}
-
-func (m *mockValidatingSubquery) ValidateForSubquery() error { return m.validateErr }
-
-func TestValidateSubqueryReturnsEarlyOnValidatorInterfaceSuccess(t *testing.T) {
-	subquery := &mockValidatingSubquery{validateErr: nil}
-	require.NoError(t, ValidateSubquery(subquery))
-}
-
-func TestValidateSubqueryWrapsValidatorInterfaceError(t *testing.T) {
-	subquery := &mockValidatingSubquery{validateErr: errors.New("missing FROM clause")}
-	err := ValidateSubquery(subquery)
-	require.Error(t, err)
-	assert.True(t, errors.Is(err, ErrInvalidSubquery),
-		"validator error must be wrapped as ErrInvalidSubquery for caller pattern-matching")
-	assert.Contains(t, err.Error(), "missing FROM clause")
-}
-
-// ---------------------------------------------------------------------------
-// TableRef.As — covers the previously-uncovered nil-receiver branch.
-// ---------------------------------------------------------------------------
-
-func TestTableRefAsReturnsErrorOnNilReceiver(t *testing.T) {
-	var t1 *TableRef // typed-nil receiver
-	_, err := t1.As("x")
-	require.Error(t, err)
-	assert.True(t, errors.Is(err, ErrNilTableRef),
-		"nil receiver must fail with ErrNilTableRef rather than panic")
 }
